+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## UNRELEASED
+
+### Added
+- Territory display fix for XML-processed geltungsbereich
+- Automatic GIS layer creation from flaechenabgrenzungsUrl in XBeteiligung messages
+- Support for MultiPolygon geometries
+
 ## v0.35 (2025-09-25)
 - fix rabbitMQ message send error for unbound procedures